+++ packages/block-library/src/file/index.js
@@ -6,7 +6,7 @@ import { includes } from 'lodash';
 /**
  * WordPress dependencies
  */
-import { __ } from '@wordpress/i18n';
+import { __, _x } from '@wordpress/i18n';
 import { createBlobURL } from '@wordpress/blob';
 import { createBlock } from '@wordpress/blocks';
 import { select } from '@wordpress/data';
@@ -23,7 +23,7 @@ export const name = 'core/file';
 export const settings = {
 	title: __( 'File' ),
 
-	description: __( 'Add a link to a file that visitors can download.' ),
+	description: __( 'Add a link to a downloadable file.' ),
 
 	icon: <SVG viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg"><Path fill="none" d="M0 0h24v24H0V0z" /><Path d="M9 6l2 2h9v10H4V6h5m1-2H4L2 6v12l2 2h16l2-2V8l-2-2h-8l-2-2z" /></SVG>,
 
@@ -39,6 +39,7 @@ export const settings = {
 			type: 'string',
 		},
 		fileName: {
+			type: 'string',
 			source: 'html',
 			selector: 'a:not([download])',
 		},
@@ -61,9 +62,10 @@ export const settings = {
 			default: true,
 		},
 		downloadButtonText: {
+			type: 'string',
 			source: 'html',
 			selector: 'a[download]',
-			default: __( 'Download' ),
+			default: _x( 'Download', 'button label' ),
 		},
 	},