+++ packages/block-library/src/heading/edit.native.js
@@ -28,6 +28,7 @@ class HeadingEdit extends Component {
 		const {
 			attributes,
 			setAttributes,
+			mergeBlocks,
 			insertBlocksAfter,
 		} = this.props;
 
@@ -58,6 +59,7 @@ class HeadingEdit extends Component {
 							content: newParaBlock.attributes.content,
 						} );
 					} }
+					onMerge={ mergeBlocks }
 					onSplit={
 						insertBlocksAfter ?
 							( before, after, ...blocks ) => {