+++ packages/blocks/src/api/registration.js
@@ -3,7 +3,7 @@
 /**
  * External dependencies
  */
-import { get, isFunction, some } from 'lodash';
+import { get, isFunction, some, reject } from 'lodash';
 
 /**
  * WordPress dependencies
@@ -174,10 +174,10 @@ export function unregisterBlockType( name ) {
 /**
  * Assigns name of block for handling non-block content.
  *
- * @param {string} name Block name.
+ * @param {string} blockName Block name.
  */
-export function setFreeformContentHandlerName( name ) {
-	dispatch( 'core/blocks' ).setFreeformFallbackBlockName( name );
+export function setFreeformContentHandlerName( blockName ) {
+	dispatch( 'core/blocks' ).setFreeformFallbackBlockName( blockName );
 }
 
 /**
@@ -193,10 +193,10 @@ export function getFreeformContentHandlerName() {
 /**
  * Assigns name of block handling unregistered block types.
  *
- * @param {string} name Block name.
+ * @param {string} blockName Block name.
  */
-export function setUnregisteredTypeHandlerName( name ) {
-	dispatch( 'core/blocks' ).setUnregisteredFallbackBlockName( name );
+export function setUnregisteredTypeHandlerName( blockName ) {
+	dispatch( 'core/blocks' ).setUnregisteredFallbackBlockName( blockName );
 }
 
 /**
@@ -340,6 +340,25 @@ export const registerBlockStyle = ( blockName, styleVariation ) => {
 				...get( settings, [ 'styles' ], [] ),
 				styleVariation,
 			],
+		};
+	} );
+};
+
+/**
+ * Unregisters a block style variation for the given block.
+ *
+ * @param {string} blockName          Name of block (example: “core/latest-posts”).
+ * @param {string} styleVariationName Name of class applied to the block.
+ */
+export const unregisterBlockStyle = ( blockName, styleVariationName ) => {
+	addFilter( 'blocks.registerBlockType', `${ blockName }/${ styleVariationName }/unregister`, ( settings, name ) => {
+		if ( blockName !== name ) {
+			return settings;
+		}
+
+		return {
+			...settings,
+			styles: reject( get( settings, [ 'styles' ], [] ), { name: styleVariationName } ),
 		};
 	} );
 };