+++ packages/editor/src/store/test/selectors.js
@@ -82,6 +82,7 @@ const {
 	isFirstMultiSelectedBlock,
 	getBlockMode,
 	isTyping,
+	isCaretWithinFormattedText,
 	getBlockInsertionPoint,
 	isBlockInsertionPointVisible,
 	isSavingPost,
@@ -89,12 +90,11 @@ const {
 	didPostSaveRequestFail,
 	getSuggestedPostFormat,
 	getEditedPostContent,
-	getNotices,
-	getReusableBlock,
-	isSavingReusableBlock,
-	isFetchingReusableBlock,
+	__experimentalGetReusableBlock: getReusableBlock,
+	__experimentalIsSavingReusableBlock: isSavingReusableBlock,
+	__experimentalIsFetchingReusableBlock: isFetchingReusableBlock,
 	isSelectionEnabled,
-	getReusableBlocks,
+	__experimentalGetReusableBlocks: getReusableBlocks,
 	getStateBeforeOptimisticTransaction,
 	isPublishingPost,
 	isPublishSidebarEnabled,
@@ -2720,6 +2720,24 @@ describe( 'selectors', () => {
 		} );
 	} );
 
+	describe( 'isCaretWithinFormattedText', () => {
+		it( 'returns true if the isCaretWithinFormattedText state is also true', () => {
+			const state = {
+				isCaretWithinFormattedText: true,
+			};
+
+			expect( isCaretWithinFormattedText( state ) ).toBe( true );
+		} );
+
+		it( 'returns false if the isCaretWithinFormattedText state is also false', () => {
+			const state = {
+				isCaretWithinFormattedText: false,
+			};
+
+			expect( isCaretWithinFormattedText( state ) ).toBe( false );
+		} );
+	} );
+
 	describe( 'isSelectionEnabled', () => {
 		it( 'should return true if selection is enable', () => {
 			const state = {
@@ -2743,6 +2761,40 @@ describe( 'selectors', () => {
 	} );
 
 	describe( 'getBlockInsertionPoint', () => {
+		it( 'should return the explicitly assigned insertion point', () => {
+			const state = {
+				currentPost: {},
+				preferences: { mode: 'visual' },
+				blockSelection: {
+					start: 'clientId2',
+					end: 'clientId2',
+				},
+				editor: {
+					present: {
+						blocksByClientId: {
+							clientId1: { clientId: 'clientId1' },
+							clientId2: { clientId: 'clientId2' },
+						},
+						blockOrder: {
+							'': [ 'clientId1' ],
+							clientId1: [ 'clientId2' ],
+							clientId2: [],
+						},
+						edits: {},
+					},
+				},
+				insertionPoint: {
+					rootClientId: undefined,
+					index: 0,
+				},
+			};
+
+			expect( getBlockInsertionPoint( state ) ).toEqual( {
+				rootClientId: undefined,
+				index: 0,
+			} );
+		} );
+
 		it( 'should return an object for the selected block', () => {
 			const state = {
 				currentPost: {},
@@ -2763,12 +2815,11 @@ describe( 'selectors', () => {
 						edits: {},
 					},
 				},
-				isInsertionPointVisible: false,
+				insertionPoint: null,
 			};
 
 			expect( getBlockInsertionPoint( state ) ).toEqual( {
 				rootClientId: undefined,
-				layout: undefined,
 				index: 1,
 			} );
 		} );
@@ -2795,42 +2846,11 @@ describe( 'selectors', () => {
 						edits: {},
 					},
 				},
-				isInsertionPointVisible: false,
+				insertionPoint: null,
 			};
 
 			expect( getBlockInsertionPoint( state ) ).toEqual( {
 				rootClientId: 'clientId1',
-				layout: undefined,
-				index: 1,
-			} );
-		} );
-
-		it( 'should return an object for the selected block with layout', () => {
-			const state = {
-				currentPost: {},
-				preferences: { mode: 'visual' },
-				blockSelection: {
-					start: 'clientId1',
-					end: 'clientId1',
-				},
-				editor: {
-					present: {
-						blocksByClientId: {
-							clientId1: { clientId: 'clientId1', attributes: { layout: 'wide' } },
-						},
-						blockOrder: {
-							'': [ 'clientId1' ],
-							clientId1: [],
-						},
-						edits: {},
-					},
-				},
-				isInsertionPointVisible: false,
-			};
-
-			expect( getBlockInsertionPoint( state ) ).toEqual( {
-				rootClientId: undefined,
-				layout: 'wide',
 				index: 1,
 			} );
 		} );
@@ -2857,12 +2877,11 @@ describe( 'selectors', () => {
 						edits: {},
 					},
 				},
-				isInsertionPointVisible: false,
+				insertionPoint: null,
 			};
 
 			expect( getBlockInsertionPoint( state ) ).toEqual( {
 				rootClientId: undefined,
-				layout: undefined,
 				index: 2,
 			} );
 		} );
@@ -2889,21 +2908,31 @@ describe( 'selectors', () => {
 						edits: {},
 					},
 				},
-				isInsertionPointVisible: false,
+				insertionPoint: null,
 			};
 
 			expect( getBlockInsertionPoint( state ) ).toEqual( {
 				rootClientId: undefined,
-				layout: undefined,
 				index: 2,
 			} );
 		} );
 	} );
 
 	describe( 'isBlockInsertionPointVisible', () => {
-		it( 'should return the value in state', () => {
+		it( 'should return false if no assigned insertion point', () => {
 			const state = {
-				isInsertionPointVisible: true,
+				insertionPoint: null,
+			};
+
+			expect( isBlockInsertionPointVisible( state ) ).toBe( false );
+		} );
+
+		it( 'should return true if assigned insertion point', () => {
+			const state = {
+				insertionPoint: {
+					rootClientId: undefined,
+					index: 5,
+				},
 			};
 
 			expect( isBlockInsertionPointVisible( state ) ).toBe( true );
@@ -3278,19 +3307,6 @@ describe( 'selectors', () => {
 			const content = getEditedPostContent( state );
 
 			expect( content ).toBe( '<!-- wp:default {\"modified\":true} /-->' );
-		} );
-	} );
-
-	describe( 'getNotices', () => {
-		it( 'should return the notices array', () => {
-			const state = {
-				notices: [
-					{ id: 'b', content: 'Post saved' },
-					{ id: 'a', content: 'Error saving' },
-				],
-			};
-
-			expect( getNotices( state ) ).toEqual( state.notices );
 		} );
 	} );