+++ packages/rich-text/src/register-format-type.js
@@ -7,18 +7,23 @@ import { isFunction } from 'lodash';
  * WordPress dependencies
  */
 import { select, dispatch } from '@wordpress/data';
-import { isValidIcon, normalizeIconObject } from '@wordpress/blocks';
 
 /**
  * Registers a new format provided a unique name and an object defining its
  * behavior.
  *
+ * @param {string} name     Format name.
  * @param {Object} settings Format settings.
  *
  * @return {?WPFormat} The format, if it has been successfully registered;
  *                     otherwise `undefined`.
  */
-export function registerFormatType( settings ) {
+export function registerFormatType( name, settings ) {
+	settings = {
+		name,
+		...settings,
+	};
+
 	if ( typeof settings.name !== 'string' ) {
 		window.console.error(
 			'Format names must be strings.'
@@ -64,16 +69,6 @@ export function registerFormatType( settings ) {
 	if ( typeof settings.title !== 'string' ) {
 		window.console.error(
 			'Format titles must be strings.'
-		);
-		return;
-	}
-
-	settings.icon = normalizeIconObject( settings.icon );
-
-	if ( ! isValidIcon( settings.icon.src ) ) {
-		window.console.error(
-			'The icon passed is invalid. ' +
-			'The icon should be a string, an element, a function, or an object following the specifications documented in https://wordpress.org/gutenberg/handbook/format-api/#icon-optional'
 		);
 		return;
 	}